Change models event methods from public to protected

+++ database/migrations/2018_02_12_000300_create_online_tracker_table.php
@@ -18,6 +18,9 @@ class CreateOnlineTrackerTable extends Migration
 
     public function down()
     {
+        if (!Schema::hasTable('igniter_onlinetracker_tracker'))
+            return;
+
         Schema::table('igniter_onlinetracker_tracker', function (Blueprint $table) {
             $table->dropColumn('session_id');
             $table->dropColumn('geoip_id');
